% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson3.R
\name{is.leap}
\alias{is.leap}
\title{Leap year function}
\source{
https://en.wikipedia.org/wiki/Leap_year#Gregorian_calendar
}
\usage{
is.leap(year)
}
\arguments{
\item{year}{the year to check. This should be an integer equal or larger than 1582}
}
\value{
TRUE if it is a leap year, FALSE otherwise
}
\description{
Function returns wether a year is a leap year.
}
\examples{
is.leap(2000)
is.leap(1900)
}
\author{
Team Maja, Gijs Peters & Simon Veen
}

